ReversedSeries

Documentation for ReversedSeries.

ReversedSeries.Reversed

ReversedSeries.ReversedFrame

Base.getindex

Base.getproperty

ReversedSeries.crossed_down

ReversedSeries.crossed_down_now

ReversedSeries.crossed_up

ReversedSeries.crossed_up_now

ReversedSeries.regular_bearish

ReversedSeries.regular_bullish

ReversedSeries.Reversed — Type

struct Reversed{T<:AbstractArray}

A::AbstractArray

This presents a reversed view of anything array-like.

Credit: @simsurace

Example

julia> a = [1, 2, 3]
julia> r = Reversed(a)
julia> r[1]
3
julia> r[3]
1
julia> a[end] == r[1]
true

ReversedSeries.ReversedFrame — Type

struct ReversedFrame{df<:DataFrames.DataFrame}

__df::DataFrames.DataFrame

__s::Dict

This presents a reversed view of an entire DataFrame.

Example

julia> rf = ReversedFrame(btcusd) # btcusd is a DataFrame
julia> rf.ts[1] # the most recent timestamp
julia> rf.ts[2] # the timestamp before that
julia> rf.c[1]  # the close at 1 corresponds with the timestamp at 1
julia> rf.o[1] == btcusd.o[end]
true

Index 1 is the origin and represents the present time.

Higher indices go back in time.

This is similar to how series are indexed in TradingView's PineScript except they use a 0-based index.

Base.getindex — Method

getindex(rf::ReversedFrame, k::Symbol) -> Any

Return a reversed view of a DataFrame column via indexing.

Base.getproperty — Method

getproperty(rf::ReversedFrame, k::Symbol) -> Any

Return a reversed view of a DataFrame column via key.

ReversedSeries.crossed_down — Method

crossed_down(a, b; i) -> Any

Did series a become less than series b at index i.

ReversedSeries.crossed_down_now — Method

crossed_down_now(a, b; i) -> Any

Is series a currently less than series b?

ReversedSeries.crossed_up — Method

crossed_up(a, b; i) -> Any

Did series a become greater than series b at index i.

ReversedSeries.crossed_up_now — Method

crossed_up_now(a, b; i) -> Any

Is series a currently greater than series b?

ReversedSeries.regular_bearish — Method

regular_bearish()

ReversedSeries.regular_bullish — Method

regular_bullish()